pyproject.toml
[tool.poetry]
name = "trend-from-qiita"
version = "0.1.0"
description = ""
authors = ["Kazuhiro-Mimaki <godblessu0405@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
tweepy = "^4.4.0"
bs4 = "^0.0.1"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
